chapter_2/simple_app/requirements.txt
django==1.7
djangorestframework==3.0.2
markdown==2.5.2
django-filter==0.9.1